Browser test cleanup
