+++ browser-test/src/support/index.ts
@@ -1,14 +1,10 @@
-import {test, expect, Page} from '@playwright/test'
-import {AxeBuilder} from '@axe-core/playwright'
-import {waitForPageJsLoad} from './wait'
-import {
-  BASE_URL,
-  LOCALSTACK_URL,
-  TEST_USER_AUTH_STRATEGY,
-  TEST_USER_LOGIN,
-  TEST_USER_PASSWORD,
-  TEST_USER_DISPLAY_NAME,
-} from './config'
+/*
+
+This barrel file is solely for backwards compatibility with existing tests.
+
+Do NOT add functions to this file.
+
+*/
 
 export {AdminQuestions} from './admin_questions'
 export {AdminPredicates} from './admin_predicates'
@@ -23,378 +19,28 @@ export {ApplicantQuestions} from './applicant_questions'
 export {ClientInformation, TIDashboard} from './ti_dashboard'
 export {clickAndWaitForModal, dismissModal, waitForPageJsLoad} from './wait'
 export {validateScreenshot} from './screenshots'
-export {normalizeElements} from './helpers'
+export {extractEmailsForRecipient, supportsEmailInspection} from './email'
+export {
+  loginAsAdmin,
+  loginAsCiviformAndProgramAdmin,
+  loginAsProgramAdmin,
+  loginAsTestUser,
+  loginAsTrustedIntermediary,
+  logout,
+  testUserDisplayName,
+} from './auth'
+export {
+  isLocalDevEnvironment,
+  dismissToast,
+  selectApplicantLanguageNorthstar,
+  disableFeatureFlag,
+  enableFeatureFlag,
+  closeWarningMessage,
+  validateAccessibility,
+  normalizeElements,
+  validateToastHidden,
+  validateToastMessage,
+  throttle,
+} from './helpers'
 
-export const isLocalDevEnvironment = () => {
-  return (
-    BASE_URL === 'http://civiform:9000' || BASE_URL === 'http://localhost:9999'
-  )
-}
-
-/**
- * Different auth strategies that are being exercised in this test. Each strategy
- * requires different logic for login (which fields to fill and button to click on
- * login page) and logout (some logout flows require confirmation).
- */
-export enum AuthStrategy {
-  FAKE_OIDC = 'fake-oidc',
-  AWS_STAGING = 'aws-staging',
-  SEATTLE_STAGING = 'seattle-staging',
-}
-
-/** True when the test environment is hermetic i.e. not a durable staging deployment. */
-export const isHermeticTestEnvironment = () =>
-  TEST_USER_AUTH_STRATEGY === AuthStrategy.FAKE_OIDC
-
-export const dismissToast = async (page: Page) => {
-  await page.locator('#toast-container div:text("x")').click()
-  await waitForPageJsLoad(page)
-}
-
-export const logout = async (page: Page, closeToast = true) => {
-  await test.step('Logout', async () => {
-    await page.click('#logout-button')
-    // If the user logged in through OIDC previously - during logout they are
-    // redirected to dev-oidc:PORT/session/end page. There they need to confirm
-    // logout.
-    if (page.url().match('dev-oidc.*/session/end')) {
-      const pageContent = await page.textContent('html')
-      if (pageContent!.includes('Do you want to sign-out from')) {
-        // OIDC central provider confirmation page
-        await page.click('button:has-text("Yes")')
-      }
-    }
-
-    // Logout is handled by the play framework so it doesn't land on a
-    // page with civiform js where we should waitForPageJsLoad. Because
-    // the process goes through a sequence of redirects we need to wait
-    // for the final destination URL (the programs index page), to make tests reliable.
-    await page.waitForURL('**/programs')
-    await validateToastMessage(page, 'Your session has ended.')
-    if (closeToast) await dismissToast(page)
-  })
-}
-
-export const loginAsAdmin = async (page: Page) => {
-  await test.step('Login as Civiform Admin', async () => {
-    await page.click('#debug-content-modal-button')
-    await page.click('#admin')
-    await waitForPageJsLoad(page)
-  })
-}
-
-export const loginAsProgramAdmin = async (page: Page) => {
-  await test.step('Login as Program Admin', async () => {
-    await page.click('#debug-content-modal-button')
-    await page.click('#program-admin')
-    await waitForPageJsLoad(page)
-  })
-}
-
-export const loginAsCiviformAndProgramAdmin = async (page: Page) => {
-  await test.step('Login as Civiform and Program Admin', async () => {
-    await page.click('#debug-content-modal-button')
-    await page.click('#dual-admin')
-    await waitForPageJsLoad(page)
-  })
-}
-
-export const loginAsTrustedIntermediary = async (page: Page) => {
-  await test.step('Login as Trusted Intermediary', async () => {
-    await page.click('#debug-content-modal-button')
-    await page.click('#trusted-intermediary')
-    await waitForPageJsLoad(page)
-  })
-}
-
-/**
- * Logs in via an auth provider.
- * @param loginButton Selector of a button on current page that starts auth
- *     login. Normally it's "Log in" button on main page, but in some cases
- *     login can be initiated from different pages, for example after program
- *     submission.
- */
-export const loginAsTestUser = async (
-  page: Page,
-  loginButton = 'a:has-text("Log in")',
-  isTi = false,
-  displayName: string = '',
-) => {
-  await test.step('Login as Test User', async () => {
-    switch (TEST_USER_AUTH_STRATEGY) {
-      case AuthStrategy.FAKE_OIDC:
-        await loginAsTestUserFakeOidc(page, loginButton, isTi)
-        break
-      case AuthStrategy.AWS_STAGING:
-        await loginAsTestUserAwsStaging(page, loginButton, isTi)
-        break
-      case AuthStrategy.SEATTLE_STAGING:
-        await loginAsTestUserSeattleStaging(page, loginButton)
-        break
-      default:
-        throw new Error(
-          `Unrecognized or unset TEST_USER_AUTH_STRATEGY environment variable of '${TEST_USER_AUTH_STRATEGY}'`,
-        )
-    }
-    await waitForPageJsLoad(page)
-    if (displayName === '') {
-      displayName = testUserDisplayName()
-    }
-    await page.waitForSelector(`:has-text("Logged in as ${displayName}")`)
-  })
-}
-
-async function loginAsTestUserSeattleStaging(page: Page, loginButton: string) {
-  await page.click(loginButton)
-  // Wait for the IDCS login page to make sure we've followed all redirects.
-  // If running this against a site with a real IDCS (i.e. staging) and this
-  // test fails with a timeout try re-running the tests. Sometimes there are
-  // just transient network hiccups that will pass on a second run.
-  // In short: If using a real IDCS retry test if this has a timeout failure.
-  await page.waitForURL('**/#/login*')
-  await page.fill('input[name=userName]', TEST_USER_LOGIN)
-  await page.fill('input[name=password]', TEST_USER_PASSWORD)
-  await page.click('button:has-text("Login"):not([disabled])')
-  // eslint-disable-next-line playwright/no-wait-for-navigation
-  await page.waitForNavigation({waitUntil: 'networkidle'})
-}
-
-async function loginAsTestUserAwsStaging(
-  page: Page,
-  loginButton: string,
-  isTi: boolean,
-) {
-  await Promise.all([
-    // eslint-disable-next-line playwright/no-networkidle
-    page.waitForURL('**/u/login*', {waitUntil: 'networkidle'}),
-    page.click(loginButton),
-  ])
-
-  await page.fill('input[name=username]', TEST_USER_LOGIN)
-  await page.fill('input[name=password]', TEST_USER_PASSWORD)
-  await Promise.all([
-    page.waitForURL(isTi ? '**/admin/**' : /.*\/programs.*/, {
-      // eslint-disable-next-line playwright/no-networkidle
-      waitUntil: 'networkidle',
-    }),
-    // Auth0 has an additional hidden "Continue" button that does nothing for some reason
-    // getByRole selects items by their accessible name, so it only selects the visible button
-    page.getByRole('button', {name: 'Continue', exact: true}).click(),
-  ])
-}
-
-async function loginAsTestUserFakeOidc(
-  page: Page,
-  loginButton: string,
-  isTi: boolean,
-) {
-  await page.click(loginButton)
-  await page.waitForURL('**/interaction/*')
-
-  // If the user has previously signed in to the provider, a prompt is shown
-  // to reauthorize rather than sign-in. In this case, click "Continue" instead
-  // and skip filling out any login information. If we want to support logging
-  // in as multiple users, this will need to be adjusted.
-  const pageText = await page.innerText('html')
-  if (
-    pageText.includes(
-      'the client is asking you to confirm previously given authorization',
-    )
-  ) {
-    throw new Error(
-      'Unexpected reauthorization page. Central logout should fully logout user.',
-    )
-  }
-
-  await page.fill('input[name=login]', TEST_USER_LOGIN)
-  await page.fill('input[name=password]', TEST_USER_PASSWORD)
-
-  await page.click('button:has-text("Sign-in"):not([disabled])')
-  await page.waitForURL('**/interaction/*')
-
-  // A screen is shown prompting the user to authorize a set of scopes.
-  // This screen is skipped if the user has already logged in once.
-  await page.click('button:has-text("Continue")')
-  await page.waitForURL(isTi ? '**/admin/**' : /\/programs.*/)
-}
-
-export const testUserDisplayName = () => {
-  if (!TEST_USER_DISPLAY_NAME) {
-    throw new Error(
-      'Empty or unset TEST_USER_DISPLAY_NAME environment variable',
-    )
-  }
-  return TEST_USER_DISPLAY_NAME
-}
-
-export const supportsEmailInspection = () => {
-  return TEST_USER_AUTH_STRATEGY === 'fake-oidc'
-}
-
-/**
- * The option to select a language is shown in the header bar as a dropdown. This helper method selects the given language from the dropdown.
- */
-export const selectApplicantLanguage = async (page: Page, language: string) => {
-  await test.step('Set applicant language from header dropdown', async () => {
-    await page.click('#select-language')
-    await page.selectOption('#select-language', {label: language})
-
-    await waitForPageJsLoad(page)
-  })
-}
-
-export const selectApplicantLanguageNorthstar = async (
-  page: Page,
-  languageCode: string,
-) => {
-  await test.step('Set applicant language from header dropdown', async () => {
-    await page.click('#select-language-menu')
-
-    await page.click(`#select-language-${languageCode}`)
-
-    await waitForPageJsLoad(page)
-  })
-}
-
-export const disableFeatureFlag = async (page: Page, flag: string) => {
-  await test.step(`Disable feature flag: ${flag}`, async () => {
-    const response = await page.goto(`/dev/feature/${flag}/disable`)
-    expect(response?.status(), {
-      message: `Could not disable feature flag '${flag}'. Make sure the flag exists.`,
-    }).toBe(200)
-  })
-}
-
-export const enableFeatureFlag = async (page: Page, flag: string) => {
-  await test.step(`Enable feature flag: ${flag}`, async () => {
-    const response = await page.goto(`/dev/feature/${flag}/enable`)
-    expect(response?.status(), {
-      message: `Could not enable feature flag '${flag}'. Make sure the flag exists.`,
-    }).toBe(200)
-  })
-}
-
-/**
- * Close the warning toast message if it is showing, otherwise the element may be in
- * the way when trying to click on various top nav elements.
- * @param {Page} page Playwright page to operate against
- */
-export const closeWarningMessage = async (page: Page) => {
-  const warningMessageLocator = page.locator('#warning-message-dismiss')
-
-  if (await warningMessageLocator.isVisible()) {
-    await warningMessageLocator.click()
-  }
-}
-
-/**
- * Run accessibility tests using axe accessibility testing engine
- * @param {Page} page Playwright page to operate against
- */
-export const validateAccessibility = async (page: Page) => {
-  await test.step(
-    'Validate accessiblity',
-    async () => {
-      const results = await new AxeBuilder({page}).analyze()
-      const errorMessage = `Found ${results.violations.length} axe accessibility violations\nOn page: ${page.url()}`
-      expect(results.violations, errorMessage).toEqual([])
-    },
-    {box: true},
-  )
-}
-
-/**
- * Check if the toast message contains the expected value
- * @param {Page} page Playwright page to operate against
- * @param {string} value Text to look for within the toast message
- */
-export const validateToastMessage = async (page: Page, value: string) => {
-  await test.step(
-    'Validate toast message',
-    async () => {
-      const toastMessages = await page.innerText('#toast-container')
-      expect(toastMessages).toContain(value)
-    },
-    {box: true},
-  )
-}
-
-export const validateToastHidden = async (page: Page) => {
-  await test.step(
-    'Validate toast hidden',
-    async () => {
-      await expect(page.locator('#toast-container')).toBeHidden()
-    },
-    {box: true},
-  )
-}
-
-type LocalstackSesResponse = {
-  messages: LocalstackSesEmail[]
-}
-type LocalstackSesEmail = {
-  Body: {
-    html_part: string | null
-    text_part: string | null
-  }
-  Destination: {
-    ToAddresses: string[]
-  }
-  Source: string
-  Subject: string
-}
-
-/**
- * Queries the emails that have been sent for a given recipient. This method requires that tests
- * run in an environment that uses localstack since it captures the emails sent using SES and makes
- * them available at a well-known endpoint). An error is thrown when the method is called from an
- * environment that does not use localstack. The supportsEmailInspection method can be used to
- * determine if the environment supports sending emails.
- */
-export const extractEmailsForRecipient = async function (
-  page: Page,
-  recipientEmail: string,
-): Promise<LocalstackSesEmail[]> {
-  if (!supportsEmailInspection()) {
-    throw new Error('Unsupported call to extractEmailsForRecipient')
-  }
-  const originalPageUrl = page.url()
-  await page.goto(`${LOCALSTACK_URL}/_aws/ses`)
-  const responseJson = JSON.parse(
-    await page.innerText('body'),
-  ) as LocalstackSesResponse
-
-  const allEmails = responseJson.messages
-  const filteredEmails = allEmails.filter((email) => {
-    return email.Destination.ToAddresses.includes(recipientEmail)
-  })
-
-  await page.goto(originalPageUrl)
-  return filteredEmails
-}
-
-export const expectEnabled = async (page: Page, locator: string) => {
-  expect(await page.getAttribute(locator, 'disabled')).toBeNull()
-}
-
-export const expectDisabled = async (page: Page, locator: string) => {
-  expect(await page.getAttribute(locator, 'disabled')).not.toBeNull()
-}
-
-/**
- * This can be used to simulate slow networks to aid in debugging flaky tests. Its use *should NOT* be
- * committed into the codebase as a permanent fix to anything.
- *
- * This works by modifying the network requests of routes and adding a timeout to help extend the load
- * time of pages. Place this at the beginning of a specific test or a beforeEach call. Playwright currently
- * does not have any builtin throttling capabilities and this is the least invasive option.
- *
- * @param page Playwright page
- * @param {number} timeout in milliseconds
- */
-export const throttle = async (page: Page, timeout: number = 100) => {
-  await page.route('**/*', async (route) => {
-    await new Promise((f) => setTimeout(f, timeout))
-    await route.continue()
-  })
-}
+// Do NOT add functions to this file.
